Introduce ProcessPromise interface for the PHPStan Pro worker
ProcessPromise becomes an interface describing the fixer:worker process
as seen by FixerApplication; the existing react/child-process
implementation is renamed to SpawnedProcessPromise. FixerApplication
creates it through a factory method. Sole implementation for now — a
pcntl_fork() based one follows.
